pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "mcp-memos"
version = "0.1.0"
description = "A MCP server for interacting with Memos."
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "grpcio>=1.76",
    "mcp>=1.19",
    "pydantic>=2.12.3",
    "requests>=2.32.5",
    "rich-click>=1.9.4",
]

scripts.mcp-memos = "mcp_memos.cli:main"

[dependency-groups]
dev = [
    "googleapis-common-protos>=1.71",
    "grpcio-tools>=1.76",
    "mypy>=1.18.2",
    "pre-commit>=4.1",
    "pre-commit-uv>=4.1.4",
    "pytest>=8.3.5",
    "types-grpcio>=1.0.0.20251009",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = [ "src/mcp_memos/api" ]

[tool.pyproject-fmt]
indent = 4

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "mcp_memos.api.v1.*",
]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "google.*",
]
ignore_missing_imports = true
